# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/leads/{workspaceId}": {
      "post": {
        "operationId": "LeadsWorkspaceController_create",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "description": "ID of the workspace",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create lead",
        "tags": [
          "Leads"
        ]
      },
      "get": {
        "operationId": "LeadsWorkspaceController_findAllFromWorkspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "description": "ID of the workspace",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of leads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{workspaceId}/me": {
      "get": {
        "operationId": "LeadsWorkspaceController_findAllMineFromWorkspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "description": "ID of the workspace",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{workspaceId}/{id}": {
      "get": {
        "operationId": "LeadsWorkspaceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "description": "ID of the workspace",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lead by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Leads"
        ]
      },
      "patch": {
        "operationId": "LeadsWorkspaceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "description": "ID of the workspace",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update lead",
        "tags": [
          "Leads"
        ]
      },
      "delete": {
        "operationId": "LeadsWorkspaceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "description": "ID of the workspace",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{workspaceId}/{id}/status": {
      "patch": {
        "operationId": "LeadsWorkspaceController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "description": "ID of the workspace",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads/{workspaceId}/{id}/proposal": {
      "post": {
        "operationId": "LeadsWorkspaceController_sendProposal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "description": "ID of the workspace",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendProposalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messageId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send cooperation proposal via SMTP",
        "tags": [
          "Leads"
        ]
      }
    },
    "/api/leads": {
      "get": {
        "operationId": "LeadsController_getAllLeads",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of leads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Leads"
        ]
      }
    },
...
    "/api/workspaces": {
      "get": {
        "operationId": "WorkspaceController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseWorkspaceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      },
      "post": {
        "operationId": "WorkspaceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseFullWorkspaceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspaces/me": {
      "get": {
        "operationId": "WorkspaceController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseWorkspaceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspaces/{workspaceId}": {
      "get": {
        "operationId": "WorkspaceController_getWorkspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWorkspaceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
...
    "/api/workspaces/{id}": {
      "delete": {
        "operationId": "WorkspaceController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    }
  },
  "info": {
    "title": "Best CRM API",
    "description": "API documentation for Best CRM backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateLeadDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "lead@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1-555-1234"
          },
          "site": {
            "type": "string",
            "example": "example.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "comment": {
            "type": "string",
            "example": "Interested in product X"
          },
          "companyTypeId": {
            "type": "string",
            "example": "uuid-of-company-type"
          }
        },
        "required": [
          "email"
        ]
      },
      "LeadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid"
          },
          "email": {
            "type": "string",
            "example": "lead@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1-555-1234"
          },
          "site": {
            "type": "string",
            "example": "example.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "comment": {
            "type": "string",
            "example": "Interested in product X"
          },
          "status": {
            "type": "number",
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "callType": {
            "type": "number",
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "companyType": {
            "type": "object",
            "example": {
              "id": "uuid",
              "code": "b2b",
              "name": "B2B"
            },
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2025-09-18T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2025-09-18T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateLeadDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "lead@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1-555-1234"
          },
          "site": {
            "type": "string",
            "example": "example.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "comment": {
            "type": "string",
            "example": "Interested in product X"
          },
          "companyTypeId": {
            "type": "string",
            "example": "uuid-of-company-type"
          }
        },
        "required": [
          "email"
        ]
      },
      "SendProposalDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "example": "Коммерческое предложение",
            "nullable": true
          }
        },
        "required": [
          "subject"
        ]
      },
...
      "ResponseWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Test Workspace"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ]
      },
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Test Workspace"
          }
        },
        "required": [
          "name"
        ]
      },
      "ResponseFullWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Test Workspace"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      }
    }
  }
}